// ui/src/types/diary.ts
export type Ship = string;
// ...
export type NoteContent = Verse[];

export interface NoteEssay {
  title: string;
  image: string;
  content: NoteContent;
  author: Ship;
  sent: number;
}

export interface DiaryNote {
  type: 'note';
  seal: NoteSeal;
  essay: NoteEssay;
}

export interface DiaryOutline extends NoteEssay {
  type: 'outline';
  quipCount: number;
  quippers: Ship[];
}